Move the color picker listener to the right namespace
