+++ build.default.properties
@@ -18,7 +18,8 @@ phpcpd.excluded=DataDefinition/Palette/Builder/Event \
                 Event \
 
 # Excluded public sources from phpcs.
-phpcs.excluded=src/Resources/public \
+phpcs.excluded=deprecated-autoload.php \
+               src/Resources/public \
                src/Resources/contao
 
 autoload-validation.customflags=--add-autoloader=autoload-validation-hack.php
